Consider "demo" plugin to be a system one
OCPN_BUILD_SAMPLE build option is disabled by default and if somebody
enables it, it really should make the sample plugin available, but this
wasn't the case before because it was installed into the system plugin
directory but not considered a system plugin and so was never loaded
from there.

Fix this by adding it to the system plugin list. Ideally this should be
only done when OCPN_BUILD_SAMPLE is on, but for now just do it always,
as it doesn't do any real harm even if it's off.