First stab on package generation (#4070)
The ci/generic-build-debian.sh now is capable of either making a plain
build or to create a debian package using the setup from packaging
branches

The used path depends on circleci/config.yml. If there is a
PACKAGE_BRANCH environment variable defined for the build a Debian
package will be built

As of this commit, the bookworm builds does build a Debian
package, all others just make a plain build.

Closes: #4070