Alternative bad plugin blocking - #4121
Disable attempts to load a plugin which once has failed. Implemented
by creating a stamp file when the load process starts. After a
successful load the stamp file is removed.

If the stamp file exists when the plugin is about to be loaded the
load process aborts.

Updating or removing the plugin clears the stamp file, making a new
load attempt possible.