tile_queue: Plug concurrency window
Current code has a small window between a thread being released
by the semaphore and, in next step, acquiring the lock. In this
window anither thread might grab the lock and thus create a
crash later.

Use a condition variable instead of a semaphore. This keeps
the lock while testing it making things consistent.