Avoid wxThread, fix tile dtor
Using wxThread comes with some drawbacks, notably that the wxWidgets
clean up processes might get involved when terminating the thread.

Use a std::thread instaed. This is much more flexible, and the clean up
is just the C++ dtor.

In the tile dtor, mark fre'ed 6GL resources as such to avoid multiple
free() on the same block causing a crash.