Use paths from manual_pi for entries
The paths to the manual_pi off-line manual is now imported from a json
file provided by manual_pi. This means things will continue to work
even if the paths are modified in a new release.

The urls to the on-line manual are still hardcoded. The only way to work
around the inevitable failure if the links are modified would be to
announce the entrypoints to some site which just returned html redirects
to correct location.