NMEA debug window: Hide until OK pressed (#3923)
The nmea window is a non-modal frame. When launched from the options
modal dialog this dialog steals the focus and while the NMEA frame is
visible it's in many ways non-functional.

Hide the window until the modal dialog is closed, update the label
for "Show NMEA Debug Window" to reflect the new beaviour"