windows fix: Drop Debug root context submenu
The windows theming mechanism is inferior to Linux and MacOS. One
result is that submenus don't inherit for example background colour
from the theme. Furthermore, wxWidgets has no way to set background
colour for a submenu.

Drop the Debug submenu for now and add "Start NMEA debug log" directly
to the root context menu, adding background/foreground on Windows to
compensate for the strange theming.