fix: cmake if block always evaluating to true
Fixed the if statement so that the catkin block does not always
evaluate to true. The ON and OFF settings are used as True or False
in CMake and therefore that block did not work as the author intended.

This is a big deal if you already have the catkin package installed,
but for new developers, your code will fail to build even though
the catkin unit tests are supposed to be off by default.