CmakeLists: make sure libudev is reflected in config.h - #4605
Libudev must be searched for by global CMakeLists since this is the
one which generates config.h. #4605 is caused by global CMakeLists not
setting HAVE_LIBUDEV in config.h.

This is a mess -- the code for locating libudev is duplicated in
CMakeLists.txt and model/CMakeLists.txt. This needs to be sorted out
in the 5.14 cycle.

Closes: #4605