data monitor et. al.: Use "handled" instead of "received" - #4644
The term "received" for a message recognized by the multiplexer is
x confusing. Problem is that "received" has a more general meaning.
In particular, an "input" event reflects a received message which
makes it a bad idea to use "received" for something else.

Use the term "handled" instead for a message handled by the multiplexer,
i.e., a message which has a corresponding listener handling it.

Closes: #4644