gl_texture_manager: clean up fishy loop - #4822
As discussed in #4821 the first loop affected by this patch could
be subject to related to an invalidated iterator.

The second loop is void since the first loop removes any possible
match in the second.

Rewrite first loop using the standard remove-erase idiom and remove
the second. While on it, clean up a related debug statement.