ocpn_plugin.h: api 121: Fix factory and nested Route linkage
The factory method GetHostApi() is the only method actually linked
to the plugin. For this to work on Windows a
__declspec(dllimport/dllexport) is needed. Add this.

All other functions are virtual and part of the vtable returned by
GetHostApi() and does not need any __declspec.

The nested class Route is returned as a unique_ptr<Route>. For this
to work the dtor/ctor needs to known by the calling code and must
thus be defined in the header. Add explicit definitions and remove
in .cpp file.