ocpn_app: Enhance exception handling
Most if not all wxWidgets libraries are compiled with wxUSE_STL
unset. For unknown reasons this affects how unhandled exceptions
are handled, and the exception  what() string is not presented in
any way.

Override the default OnUnhandledException(), basically doing the
same as the original code, but without the #ifdef wxUSE_STL
condition. Net result is that the exception string is displayed.

This should be safe, all compilers used in the project
should be aware of std::exception.